Add ejs packages, reconfigure server.js, and convert html to ejs
