+++ views/index.ejs
@@ -1,0 +1,30 @@
+<div class="form-header">
+  <h2>Stevens 150th Anniversary Nomination</h2>
+</div>
+<form method="post" action="/" novalidate>
+  <div class="form-field">
+    <label for="nominee full name">Nominee Full Name</label>
+    <input
+      class="input"
+      id="nominee-name"
+      name="nominee-full-name"
+      type="text"
+      alt="Nominee full name"
+      placeholder="Nominee Name"
+    />
+  </div>
+  <div class="form-field">
+    <label for="explanation">Explanation</label>
+    <textarea
+      class="input"
+      id="explanation"
+      name="explanation"
+      rows="4"
+      alt="Explanation for why this nominee"
+      placeholder="Please provide a brief explanation of why you would like to nominate this person."
+    ></textarea>
+  </div>
+  <div class="form-actions">
+    <button class="btn" type="submit">Submit</button>
+  </div>
+</form>
